bugfix/#58 fixed export envp leak + tests
