+++ tests/tests.txt
@@ -8,7 +8,7 @@ cat Makefile || wc
 <<EOF<<EOF2<<EOF3 << cat							[!] SEG
 cat tasks.md > grep "a"								[!] Valgrind LEAK
 cat << EOF > output.txt								[!] Valgrind LEAK
-echo hello											[!] Valgrind LEAK
+echo mate > hello									[!] Valgrind LEAK
 'EOF' / "..."										[!] Valgrind LEAK
 grep aa << EOF
 echo hello to this >> input.txt 
@@ -63,8 +63,10 @@ echo e"$ch"o										[!] Valgrind LEAK
 
 
 === BUILTINS ===
-export A=1 B=2 C=3 D=4 E=5 F=6 G=7 H=8 I=9 J=10
-export A=200										[!] ERROR creates new
+export A=1 B=2 C=3 D=4 E=5 F=6 G=7 H=8 I=9 J=10		[!] ERROR (only once when multiple is called. second time is none)
+export =
+export @
+export 1											
 echo "hello $A $A"
 export A=1 NEW=you B=2
 echo "Hello $NEW, copy $A $B"
